Update output functions
